Refactored derive macro in preparation for delegate_to_methods
