+++ ambassador/Cargo.toml
@@ -21,7 +21,6 @@ syn = { version = "1.0.8", features = ["full", "extra-traits", "visit"] }
 quote = "1.0.2"
 proc-macro2 = "1.0.6"
 itertools = "0.10.3"
-bytemuck = {version = "1.7.3", features = ["derive"]}
 
 [dev-dependencies]
 compiletest_rs = { version = "0.3.25", features = [ "stable" ]}
